+++ sdk-lib/src/androidTest/java/cash/z/ecc/fixture/FakeRustBackend.kt
@@ -127,6 +127,10 @@ internal class FakeRustBackend(
         error("Intentionally not implemented yet.")
     }
 
+    override suspend fun redactPcztForSigner(pczt: ByteArray): ByteArray {
+        error("Intentionally not implemented yet.")
+    }
+
     override suspend fun addProofsToPczt(pczt: ByteArray): ByteArray {
         error("Intentionally not implemented yet.")
     }