+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackend.kt
@@ -87,6 +87,15 @@ internal interface TypesafeBackend {
     ): Pczt
 
     /**
+     * Redacts information from the given PCZT that is unnecessary for the Signer role.
+     *
+     * @return the updated PCZT in its serialized format.
+     *
+     * @throws RuntimeException as a common indicator of the operation failure
+     */
+    suspend fun redactPcztForSigner(pczt: Pczt): Pczt
+
+    /**
      * Adds proofs to the given PCZT.
      *
      * @return the updated PCZT in its serialized format.