+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/OutboundTransactionManager.kt
@@ -100,6 +100,8 @@ internal interface OutboundTransactionManager {
         proposal: Proposal
     ): Pczt
 
+    suspend fun redactPcztForSigner(pczt: Pczt): Pczt
+
     suspend fun addProofsToPczt(pczt: Pczt): Pczt
 
     suspend fun extractAndStoreTxFromPczt(