+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/OutboundTransactionManagerImpl.kt
@@ -110,6 +110,8 @@ internal class OutboundTransactionManagerImpl(
         proposal: Proposal
     ) = encoder.createPcztFromProposal(accountUuid, proposal)
 
+    override suspend fun redactPcztForSigner(pczt: Pczt): Pczt = encoder.redactPcztForSigner(pczt)
+
     override suspend fun addProofsToPczt(pczt: Pczt) = encoder.addProofsToPczt(pczt)
 
     override suspend fun extractAndStoreTxFromPczt(