+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/TransactionEncoder.kt
@@ -99,6 +99,8 @@ internal interface TransactionEncoder {
         proposal: Proposal
     ): Pczt
 
+    suspend fun redactPcztForSigner(pczt: Pczt): Pczt
+
     suspend fun addProofsToPczt(pczt: Pczt): Pczt
 
     suspend fun extractAndStoreTxFromPczt(