+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/TransactionEncoderImpl.kt
@@ -165,6 +165,18 @@ internal class TransactionEncoderImpl(
         }
     }
 
+    override suspend fun redactPcztForSigner(pczt: Pczt): Pczt {
+        return runCatching {
+            backend.redactPcztForSigner(pczt = pczt)
+        }.onSuccess {
+            Twig.debug { "Result of redactPcztForSigner: $it" }
+        }.onFailure {
+            Twig.error(it) { "Caught exception while redacting PCZT for Signer." }
+        }.getOrElse {
+            throw PcztException.RedactPcztForSignerException(it.message, it.cause)
+        }
+    }
+
     override suspend fun addProofsToPczt(pczt: Pczt): Pczt {
         return runCatching {
             saplingParamTool.ensureParams(saplingParamTool.properties.paramsDirectory)