+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `Synchronizer.getTransactionsByMemoSubstring()` has been added
+- `Synchronizer.redactPcztForSigner`
 - `TransactionId` object has been added and used instead of `FirstClassByteArray` in `TransactionOverview` and 
   `PendingTransaction` model classes